Update part5b.md : sentence clarification
The exercise 5.10 does not specify whether it's arranging the blogs in ascending order or descending. This exercise is connected to the last exercise 5.23 where it's explicitly said to sort in descending order.
Someone may have sorted in ascending order to  only discover later that the tests should be in descending and come back to change his code.